backend/requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
Flask-WTF==1.2.1
PyOTP==2.9.0
qrcode==7.4.2
python-dotenv==1.0.0
email-validator==2.1.0
cryptography==41.0.7
